DB of registered users modularized, so is more maintainable. The app code gets cleaner, smaller and more readable.
Adding logger setup, so script can log success or errors, when on DEBUG level/mode. In order the keep transparent as original functionality, the logger level is set to "info". If this value is changed to "debug", you can see the logs. This can be modified on ./config/logger-conf.js. This logging functionality is reusable on any service or app file.