updatePassword changes
Adding "break" in loop, so if the user is found, the loop stops and doesn't keep searching.

Using userExists method, so this functionality gets modular and so, reusable for the whole login-service. Besides, updatePassword code gets smaller and modular.

Using checkPassword & renewPassword methods in order to be more modular and reusable.